More renaming and deleting unused files
