+++ src/helpers/scripts.js
